version 0.2.02 update

+++ recipes/vsnp/build.sh
@@ -2,4 +2,8 @@
 
 mkdir -p $PREFIX/bin
 mv bin/*.py $PREFIX/bin
-mv dependencies $PREFIX
+
+#prevent clobbering user's reference_options_paths.txt
+if [[ ! -f $PREFIX/dependencies/reference_options_paths.txt ]]; then
+    mv dependencies $PREFIX # if text file does not exists add dependencies
+fi
